Fixing PRs Resource Content for Chrome Extension
