+++ packages/chrome-extension/src/app/components/single/SingleEditorApp.tsx
@@ -98,7 +98,8 @@ export function SingleEditorApp(props: {
         value={{
           onEditorReady: () => setTextModeEnabled(true),
           fullscreen: fullscreen,
-          textMode: textMode
+          textMode: textMode,
+          isOriginal: true
         }}
       >
         {!fullscreen && (
